Module ps_snippets::ds::queue_min

A minimum-in-queue data structure, commonly called deque DP trick.

Notably, all operations take amortized constant time.

If you need a maximum-in-queue data structure, wrap it with std::cmp::Reverse.

T must implement PartialOrd.

Initialization

new() initializes an empty queue.

Basic Methods

len() returns the length of the queue.

is_empty() returns true iff the queue is empty.

Update

push(v) pushes v into the queue.

pop() pops from the queue and returns true if the queue is not empty; otherwise returns false.

Do not be confused with the internal data structure used to implement this, especially if you know how deque DP trick works; len and pop are concerned with its abstract representation, not its implementation. Refer to the example below for further clarification.

Query

get() returns the minimum value in the queue, if any; otherwise returns None.

most_recent() returns the most recent value inserted into the queue.

Example

use ps_snippets::ds::queue_min::*;
 
let mut q = QueueMin::new(); // []
q.push(1);
q.push(3);
q.push(2); // [1, 3, 2]
assert_eq!(q.most_recent(), Some(&2));
assert_eq!(q.get(), Some(&1));
assert!(q.pop()); // [3, 2]
assert_eq!(q.get(), Some(&2));

Practice Problems

BOJ 11003 최솟값 찾기

BOJ 10129 작은 새

Re-exports

pub use queue_min_mod::QueueMin;

Modules

queue_min_mod