Module ps_snippets::ds::segtree_lazy

A lazy segment tree.

To use it, first define a zero-sized type and implement the following LazyMonoid trait on it:

type V: the type of the node value.

type Lazy: the type of the lazy operation.

const ID: Self: the identity element of the value operation.

const LAZY_ID: Lazy: the identity element of the lazy operation.

op(lhs: Self, rhs: Self) -> Self: the value operation.

op_lazy(cur: Lazy, up: Lazy) -> Lazy: the lazy operation, where cur is the current lazy of the node and up is the lazy which is propagating from the parent node.

unlazy(v: Self, size: usize, lz: Lazy) -> Self: the updated value from v with lazy lz on a node of size size.

Both V and Lazy must implement Copy.

Let $O(t_V)$ and $O(t_L)$ be the time complexity of op and op_lazy respectively.

Initialization

new(n) constructs a lazy segtree with size n filled with T::ID. Takes $O(n)$.

from_slice(vals: &[T::V]) constructs a lazy segtree out of vals`. Takes $O(|vals| t_L)$.

Queries

update(l, r, lz: T::lazy) updates all indices in [l,r] with lz. Takes $O(t_L \log n)$.

query(l, r) returns the result of the operation over all indices in [l,r]. Takes $O((t_V + t_L) \log n)$.

Caution

If LazyMonoid does not satisfy the condition of a lazy segtree, the behavior is unspecified and meaningless.

update and query panics if 0 <= l <= r < n is not satisfied.

Examples

The following implementation supports range additon and range sum. Note that RangeFenwick (TODO) is better for this task, and this example is just for demonstration:

struct St;
 
impl LazyMonoid for St {
    type V = i64;
    type Lazy = i64;
    const ID: Self::V = 0;
    const LAZY_ID: Self::Lazy = 0;
    fn op(lhs: Self::V, rhs: Self::V) -> Self::V {
        lhs + rhs
    }
    fn op_lazy(cur: Self::Lazy, up: Self::Lazy) -> Self::Lazy {
        cur + up
    }
    fn unlazy(v: Self::V, size: usize, lz: Self::Lazy) -> Self::V {
        v + (size as i64) * lz
    }
}
 
let mut st = SegtreeLazy::<St>::from_slice(&[1,2,3,4,5]);
assert_eq!(st.query(1, 3), 2+3+4);
st.update(0, 2, 10); // [11, 12, 13, 4, 5]
assert_eq!(st.query(1, 3), 12+13+4);

Practice Problems

BOJ 10999 구간 합 구하기 2 1e6 elements, 2e4 queries

LC Range Affine Range Sum 5e5 elements, 5e5 queries

Re-exports

pub use segtree_lazy_mod::LazyMonoid;

pub use segtree_lazy_mod::SegtreeLazy;

Modules

segtree_lazy_mod