Module ps_snippets::graph::dinic

Maximum flow using Dinic’s algorithm.

Initialization & Building

new(n) initializes a graph with n vertices.

connect(s, e, cap) connects from vertex s to e with capacity cap, and returns an EdgeIndex (see below).

Solving

max_flow(src, sink) sends and returns the maximum flow from vertex src to sink. After this, indexing Dinic by EdgeIndex to get an Edge and calling used() gives the amount of flow sent along that edge.

min_cut(src, sink) sends the maximum flow from vertex src to sink, and returns the amount, along with the minimum cut between src to sink. The minimum cut is represented by the list of vertices in the src side.

Caution

The flow is not reset between successive max_flow or min_cut calls.

Time Complexity

TODO

Example

TODO

Practice Problems

TODO

Re-exports

pub use dinic_mod::Dinic;

Modules

dinic_mod